+++ scripts/deploy_functions.mjs
@@ -7,7 +7,8 @@ import { outputFile } from 'fs-extra'
 import { supa_url } from './utils.mjs'
 
 const exec = promisify(execCb)
-const folders = readdirSync('./supabase/functions').filter(file => !file.startsWith('_') && file !== 'demo')
+const folders = readdirSync('./supabase/functions')
+  .filter(file => !file.startsWith('_'))
 
 const token = process.env.SUPABASE_TOKEN || ''
 const projectRef = supa_url.split('.')[0].replace('https://', '')
@@ -22,22 +23,28 @@ try {
   await outputFile(`${homedir()}/.supabase/access-token`, token)
   // for in folders
   for (const folder of folders) {
-    const file = `./supabase/functions/${folder}/.no_verify_jwt`
+    const fileNoJWT = `./supabase/functions/${folder}/.no_verify_jwt`
+    const fileNoDeploy = `./supabase/functions/${folder}/.no_deploy`
     let command = `supabase functions deploy ${folder}`
     let no_verify_jwt = false
-    if (existsSync(file)) {
+    if (existsSync(fileNoJWT)) {
       command += ' --no-verify-jwt'
       no_verify_jwt = true
     }
-    console.log(`Upload ${folder}${no_verify_jwt ? ' no_verify_jwt' : ''}`)
-    await exec(command).then((r) => {
-      if (r.stderr) {
-        console.error(folder, r.stderr)
-        exit(1)
-      }
-      return r
-    })
-    console.log('Done ✅')
+    if (!existsSync(fileNoDeploy)) {
+      console.log(`Upload ${folder}${no_verify_jwt ? ' no_verify_jwt' : ''}`)
+      await exec(command).then((r) => {
+        if (r.stderr) {
+          console.error(folder, r.stderr)
+          exit(1)
+        }
+        return r
+      })
+      console.log('Done ✅')
+    }
+    else {
+      console.log('Ignored ⏭')
+    }
   }
 }
 catch (e) {