+++ supabase/functions/on_user_created/index.ts
@@ -1,4 +1,4 @@
-import { serve } from 'https://deno.land/std@0.140.0/http/server.ts'
+import { serve } from 'https://deno.land/std@0.145.0/http/server.ts'
 import { addDataPerson, addEventPerson, postPerson } from '../_utils/crisp.ts'
 import { createCustomer } from '../_utils/stripe.ts'
 import { supabaseAdmin } from '../_utils/supabase.ts'
@@ -21,8 +21,6 @@ serve(async (event: Request) => {
     console.log('body')
     const body = (await event.json()) as { record: definitions['users'] }
     const record = body.record
-    if (record.customer_id)
-      return sendRes()
     await supabase
       .from<definitions['apikeys']>('apikeys')
       .insert([
@@ -43,7 +41,9 @@ serve(async (event: Request) => {
         }])
     await postPerson(record.email, record.first_name, record.last_name, record.image_url ? record.image_url : undefined)
     console.log('createCustomer stripe')
-    const customer = await createCustomer(Deno.env.get('STRIPE_SECRET_KEY') || '', record.email)
+    if (record.customer_id)
+      return sendRes()
+    const customer = await createCustomer(record.email)
     const { error: dbStripeError } = await supabase
       .from<definitions['stripe_info']>('stripe_info')
       .insert({